# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Exposure Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileRole = (typeof ProfileRole)[keyof typeof ProfileRole];

export const ProfileRole = {
  admin: "admin",
  advisor: "advisor",
  client_user: "client_user",
} as const;

export interface Profile {
  id: string;
  email: string;
  fullName: string;
  role: ProfileRole;
  companyId?: string | null;
  createdAt: string;
}

export interface Company {
  id: string;
  companyName: string;
  tinOrTaxId?: string | null;
  industry?: string | null;
  country?: string | null;
  financialYear?: string | null;
  riskLevel?: string | null;
  riskScore?: number | null;
  transactionCount?: number | null;
  openFlagsCount?: number | null;
  estimatedExposure?: number | null;
  createdAt: string;
}

export interface CreateCompanyRequest {
  companyName: string;
  tinOrTaxId?: string;
  industry?: string;
  country?: string;
  financialYear?: string;
}

export interface RiskCategory {
  category: string;
  count: number;
  exposure: number;
}

export interface SeverityItem {
  severity: string;
  count: number;
}

export interface MonthlyExposure {
  month: string;
  exposure: number;
}

export interface CompanySummary {
  totalTransactions: number;
  openRisks: number;
  estimatedExposure: number;
  riskScore: number;
  riskLevel: string;
  risksByCategory: RiskCategory[];
  severityBreakdown: SeverityItem[];
  monthlyExposure: MonthlyExposure[];
}

export interface Transaction {
  id: string;
  companyId: string;
  transactionDate?: string | null;
  description?: string | null;
  reference?: string | null;
  amount?: number | null;
  currency?: string | null;
  accountCode?: string | null;
  accountCategory?: string | null;
  vendorName?: string | null;
  customerName?: string | null;
  taxType?: string | null;
  vatAmount?: number | null;
  withholdingTaxAmount?: number | null;
  transactionType?: string | null;
  createdAt: string;
}

export interface TransactionListResponse {
  data: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export type UploadTransactionsRequestTransactionsItem = {
  [key: string]: unknown;
};

export interface UploadTransactionsRequest {
  companyId: string;
  transactions: UploadTransactionsRequestTransactionsItem[];
  fileName: string;
}

export interface UploadResult {
  uploadId: string;
  rowsImported: number;
  message: string;
}

export interface TaxRiskFlag {
  id: string;
  companyId: string;
  transactionId?: string | null;
  ruleCode?: string | null;
  description?: string | null;
  severity?: string | null;
  estimatedExposure?: number | null;
  status?: string | null;
  category?: string | null;
  companyName?: string | null;
  createdAt: string;
}

export interface RiskListResponse {
  data: TaxRiskFlag[];
  total: number;
  page: number;
  limit: number;
}

export interface Report {
  id: string;
  companyId: string;
  title?: string | null;
  status?: string | null;
  summary?: string | null;
  totalExposure?: number | null;
  highRisks?: number | null;
  mediumRisks?: number | null;
  lowRisks?: number | null;
  companyName?: string | null;
  createdAt: string;
}

export interface GenerateReportRequest {
  companyId: string;
  title?: string;
}

export interface DashboardStats {
  totalClients: number;
  totalTransactions: number;
  openFlags: number;
  estimatedExposure: number;
  highRiskCompanies: number;
  recentUploads: number;
}

export interface FlagsOverTime {
  month: string;
  flags: number;
}

export interface DashboardCharts {
  riskByCategory: RiskCategory[];
  severityBreakdown: SeverityItem[];
  monthlyExposure: MonthlyExposure[];
  flagsOverTime: FlagsOverTime[];
}

export interface Upload {
  id: string;
  companyId: string;
  fileName?: string | null;
  rowCount?: number | null;
  status?: string | null;
  createdAt: string;
}

export type ListCompaniesParams = {
  search?: string;
  industry?: string;
  riskLevel?: string;
};

export type ListTransactionsParams = {
  companyId?: string;
  search?: string;
  taxType?: string;
  page?: number;
  limit?: number;
};

export type ListRisksParams = {
  companyId?: string;
  severity?: string;
  status?: string;
  page?: number;
  limit?: number;
};

export type ListReportsParams = {
  companyId?: string;
};

export type ListUploadsParams = {
  companyId?: string;
};