# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tax Exposure Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}/summary:
    get:
      operationId: getCompanySummary
      tags: [companies]
      summary: Get company analytics summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"
...
  /risks:
    get:
      operationId: listRisks
      tags: [risks]
      summary: List tax risk flags
      parameters:
        - name: companyId
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of risks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskListResponse"
  /risks/{id}/resolve:
    post:
      operationId: resolveRisk
      tags: [risks]
      summary: Resolve a risk flag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Risk resolved
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      parameters:
        - name: companyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate report for a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportRequest"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, advisor, client_user]
        companyId:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - email
        - fullName
        - role
        - createdAt
    Company:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        tinOrTaxId:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        financialYear:
          type: string
          nullable: true
        riskLevel:
          type: string
          nullable: true
        riskScore:
          type: number
          nullable: true
        transactionCount:
          type: integer
          nullable: true
        openFlagsCount:
          type: integer
          nullable: true
        estimatedExposure:
          type: number
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - companyName
        - createdAt
    CreateCompanyRequest:
      type: object
      properties:
        companyName:
          type: string
        tinOrTaxId:
          type: string
        industry:
          type: string
        country:
          type: string
        financialYear:
          type: string
      required:
        - companyName
    CompanySummary:
      type: object
      properties:
        totalTransactions:
          type: integer
        openRisks:
          type: integer
        estimatedExposure:
          type: number
        riskScore:
          type: number
        riskLevel:
          type: string
        risksByCategory:
          type: array
          items:
            $ref: "#/components/schemas/RiskCategory"
        severityBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SeverityItem"
        monthlyExposure:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyExposure"
      required:
        - totalTransactions
        - openRisks
        - estimatedExposure
        - riskScore
        - riskLevel
        - risksByCategory
        - severityBreakdown
        - monthlyExposure
...
    TaxRiskFlag:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        transactionId:
          type: string
          nullable: true
        ruleCode:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        severity:
          type: string
          nullable: true
        estimatedExposure:
          type: number
          nullable: true
        status:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - companyId
        - createdAt
    RiskListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TaxRiskFlag"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Report:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        title:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        totalExposure:
          type: number
          nullable: true
        highRisks:
          type: integer
          nullable: true
        mediumRisks:
          type: integer
          nullable: true
        lowRisks:
          type: integer
          nullable: true
        companyName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - companyId
        - createdAt
    GenerateReportRequest:
      type: object
      properties:
        companyId:
          type: string
        title:
          type: string
      required:
        - companyId
...
    RiskCategory:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        exposure:
          type: number
      required:
        - category
        - count
        - exposure
    SeverityItem:
      type: object
      properties:
        severity:
          type: string
        count:
          type: integer
      required:
        - severity
        - count
    MonthlyExposure:
      type: object
      properties:
        month:
          type: string
        exposure:
          type: number
      required:
        - month
        - exposure
...