# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Exposure Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: string;
  companyName: string;
  tinOrTaxId?: string | null;
  industry?: string | null;
  country?: string | null;
  financialYear?: string | null;
  riskLevel?: string | null;
  riskScore?: number | null;
  transactionCount?: number | null;
  openFlagsCount?: number | null;
  estimatedExposure?: number | null;
  createdAt: string;
}