# lib/api-zod/src/generated/types/companySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Exposure Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyExposure } from "./monthlyExposure";
import type { RiskCategory } from "./riskCategory";
import type { SeverityItem } from "./severityItem";

export interface CompanySummary {
  totalTransactions: number;
  openRisks: number;
  estimatedExposure: number;
  riskScore: number;
  riskLevel: string;
  risksByCategory: RiskCategory[];
  severityBreakdown: SeverityItem[];
  monthlyExposure: MonthlyExposure[];
}