# lib/api-zod/src/generated/types/monthlyExposure.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Exposure Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyExposure {
  month: string;
  exposure: number;
}