# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Exposure Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileRole } from "./profileRole";

export interface Profile {
  id: string;
  email: string;
  fullName: string;
  role: ProfileRole;
  companyId?: string | null;
  createdAt: string;
}